NEW: made share plio dialog another component
